+++ src/redturtle/volto/__init__.py
@@ -1,9 +1,7 @@
-# -*- coding: utf-8 -*-
 from plone.app.dexterity.behaviors.metadata import ICategorization
 from plone.app.dexterity.textindexer import utils
 from redturtle.volto import patches  # noqa
 from zope.i18nmessageid import MessageFactory
-
 
 _ = MessageFactory("redturtle.volto")