+++ src/redturtle/volto/adapters/stringinterp.py
@@ -1,9 +1,7 @@
-# -*- coding: utf-8 -*-
 from plone.stringinterp.adapters import BaseSubstitution
 from Products.CMFCore.interfaces import IContentish
 from redturtle.volto import _
 from zope.component import adapter
-
 
 try:
     from plone.stringinterp import _ as stringinterp_mf