+++ src/redturtle/volto/browser/find_broken_links.py
@@ -1,11 +1,10 @@
 from Acquisition import aq_base
+from io import StringIO
 from plone import api
 from plone.dexterity.utils import iterSchemata
 from plone.restapi.serializer.utils import uid_to_url
 from Products.Five import BrowserView
-from six import StringIO
 from zope.schema import getFieldsInOrder
-
 
 try:
     from collective.volto.blocksfield.field import BlocksField
@@ -16,7 +15,6 @@ except ImportError:
 
 import csv
 import logging
-
 
 logger = logging.getLogger(__name__)